Minor updates- handle id not found and fixed server error message

+++ backend/controllers/users/deleteUser.js
@@ -2,6 +2,14 @@ import { UserModel } from "../../models/user.model.js";
 
 const deleteUser = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "User not found",
+    });
+  }
+
   try {
     await UserModel.findByIdAndDelete(id);
     res.status(200).json({
@@ -10,9 +18,9 @@ const deleteUser = async (req, res) => {
     });
   } catch (error) {
     console.error(`Error: ${error.message}`);
-    res.status(404).json({
+    res.status(500).json({
       success: false,
-      message: "User not found",
+      message: "Server Error",
     });
   }
 };
